other file/requirements.txt
python-dotenv
accelerate
protobuf
scipy
sentencepiece
llama-index
langchain
langchain-experimental
PyMuPDF
sentence-transformers
safetensors
ninja
optimum
huggingface-hub
text_generation
weaviate-client
ujson
torch==2.1.0